requirements.txt
Django>=1.11.6
django-environ>=0.4.3
django-braces>=1.11.0
django-admin-bootstrapped>=2.5.7
django-filter==1.0.4
djangorestframework==3.6.4
psycopg2
coreapi
django-cors-headers==2.1.0